backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pdfplumber==0.11.0
python-docx==1.1.2
openpyxl==3.1.4
mutagen==1.47.0
Pillow==10.3.0
pydantic==2.7.1
aiofiles==23.2.1
